==> Script03_Capacity-Copy1.ipynb <==
## Importing all dependencies 
import pandas as pd
import os
import numpy as np

###Input Files
##This is the configuraiton and the only thing to change in this code whenever you rerun
cwd = os.getcwd()
Demand_Output = "28-04-2024_Demand_Automation.xlsx"
Capacity_File = os.path.join(cwd, "client_data", "Capacity.xlsx")
Line_map_data = os.path.join(cwd, "client_data", "Factory_Line_Master.xlsx")

df = pd.read_excel(Capacity_File,skiprows=1)
df.head()

dd = pd.read_excel(Demand_Output)
line_mapping = pd.read_excel(Line_map_data)

#### Filter capacity data only for the lines having production capacity excluding prep lines 

#### Filter capacity data only for the lines having production capacity excluding prep lines 
df1 = df[['Item', 'Disc', 'Weight UOM',
   'C1', 'C2', 'C3', 'C4', 'C5', 'New Line 1','G64', 'G76', 'M7', 'G4',
          'D', 'F', 'P', 'T', 'Manual NFC','New Line 2','New Line 3', 'B1', 'B2', 'Rivo',
          'Ultra Formax', 'Glazed', 'Tiromat', 'Manual KFC','New Line 4','DSI','TSM']]

df2 = pd.melt(df1, id_vars=['Item', 'Disc', 'Weight UOM'])

df2 = df2.rename(columns={
   'Item': 'Item',
   'Disc': 'Description',
   'variable': 'Line',
   'value': 'Throughput'
   })

df_base = pd.merge(df2, line_mapping, 
                   left_on=['Line'], 
                   right_on=['Line-name'], 
                   how='left')
df_base1 = df_base[['Item', 'Description', 'Line', 'Throughput', 'Factory', 'Weight UOM']]

### Filter Demand for only retained cases 
dd_filter = dd[dd['Ret_Flag'] == 1]
dd_base = pd.DataFrame({'FGSKUCode': dd_filter['FGSKUCode'].unique()})

# Map capacity against demand
Capacity_DD_Filtered = pd.merge(dd_base, df_base1, 
                                left_on=['FGSKUCode'], 
                                right_on=['Item'], 
                                how='left')
Capacity_DD_Filtered = Capacity_DD_Filtered[['FGSKUCode', 'Description', 'Line', 'Throughput', 'Factory', 'Weight UOM']]

# Convert to numeric and handle non-numeric values
Capacity_DD_Filtered['Throughput'] = pd.to_numeric(Capacity_DD_Filtered['Throughput'], errors='coerce')
Capacity_DD_Filtered['Weight UOM'] = pd.to_numeric(Capacity_DD_Filtered['Weight UOM'], errors='coerce')

# Drop rows where Throughput or Weight UOM is NaN or 0
Capacity_DD_Filtered = Capacity_DD_Filtered.dropna(subset=['Throughput', 'Weight UOM'])
Capacity_DD_Filtered = Capacity_DD_Filtered[(Capacity_DD_Filtered['Throughput'] != 0) & (Capacity_DD_Filtered['Weight UOM'] != 0)]

# Convert Throughput from Kg/Hour to Units/Hour
Capacity_DD_Filtered['Throughput'] = Capacity_DD_Filtered['Throughput'] / Capacity_DD_Filtered['Weight UOM']


Capacity_DD_Filtered['retain_flag'] = 1  # Default value is 1
Capacity_DD_Filtered.loc[(Capacity_DD_Filtered['Throughput'].isnull()) | 
                         (Capacity_DD_Filtered['Throughput'] == 0) | 
                         (Capacity_DD_Filtered['Weight UOM'].isnull()) | 
                         (Capacity_DD_Filtered['Weight UOM'] == 0), 'retain_flag'] = 0

# Handle potential division by zero or infinity
Capacity_DD_Filtered['Throughput'] = Capacity_DD_Filtered['Throughput'].replace([np.inf, -np.inf], np.nan)

if Capacity_DD_Filtered[Capacity_DD_Filtered['retain_flag'] == 1].empty:
    print('No common records found in Demand-SKU Factory Capability Mapping, please check the source data')
else:  
    ###Capacity Filtered and mapped - FG+Factory+Line - should have unique records
    Capacity_DD_Filtered = Capacity_DD_Filtered.drop_duplicates()           
    Capacity_DD_Filtered.to_excel("Capacity_DD_Filtered_w_preplines.xlsx", index=False)
    print('Data Processing for Demand-SKU Factory Capability Mapping is Successful!!. Output file name - Capacity_DD_Filtered.xlsx')

##Also did some manual changes to output to add factory name to the mapping
