==> Script06_ProductionAutomation.ipynb <==
# Importing all the dependencies
import pandas as pd
import os
import numpy as np
from openpyxl import Workbook
from openpyxl.styles import Font, NamedStyle, Border, Side, PatternFill

######Configure inputs here
# Reading the required files

##########Configuration of filenames
cwd = os.getcwd()
#input_demand = "28-04-2024_Demand_Automation.xlsx"
standard_costs = os.path.join(cwd, "client_data", "Data 3 entities.xlsx")

input_routing = os.path.join(cwd, "client_data", "routing data - 070424.xlsx")
routing_additional = os.path.join(cwd, "client_data", "Routing - Factory.xlsx")

# Read the Capacity_DD_Filtered file
df_production = pd.read_excel(r"Capacity_DD_Filtered.xlsx")
# Remove rows where Throughput is zero
df_production = df_production[df_production['Throughput'] > 0]
# Reset the index after filtering
df_production.reset_index(drop=True, inplace=True)


df_3=pd.read_excel(standard_costs,skiprows=2)

input_demand = "28-04-2024_Demand_Automation.xlsx"

#######Adding missing routing file here so that it can be included in preparation
df_missing_routing=pd.read_excel(routing_additional,sheet_name='II. FGs_missing_routing.',skiprows=2)

df_missing_routing=df_missing_routing[['Factory', 'Item Code', 'Crew\nTotal ', 'Run Rate (Primary UOM)\nFG', 'Power Usage\nFG']].rename(columns={
    'Crew\nTotal ': 'Crew Total',
    'Run Rate (Primary UOM)\nFG': 'Run Rate (Primary UOM FG)',
    'Power Usage\nFG': 'Power Usage FG'
})

df_missing_routing=df_missing_routing.rename(columns={
    'Factory':'Unit'})

conditions = [
    df_missing_routing['Unit'].isin(['NF1', 'NF2']),
    df_missing_routing['Unit'] == 'GFM',
    df_missing_routing['Unit'] == 'KPF'
]
choices = ['NFC', 'GFC', 'KFC']
df_missing_routing['FACTCode'] = np.select(conditions, choices, default=None)
df_missing_routing.to_excel('Missing_routing.xlsx',index=False)

df_routing_missing= pd.read_excel('Missing_routing.xlsx')
##This file will be used later for concatenation

# df_demand=pd.read_excel(r"01-04-2024_Demand_Automation.xlsx",sheet_name='Demand_output')
df_demand=pd.read_excel(input_demand,sheet_name='Demand_output')

df_routing_NFC1=pd.read_excel(input_routing,skiprows=4,sheet_name='NF1')
df_routing_NFC2=pd.read_excel(input_routing,skiprows=4,sheet_name='NF2')
df_routing_KFC=pd.read_excel(input_routing,skiprows=4,sheet_name='KPF')
df_routing_GFC=pd.read_excel(input_routing,skiprows=4,sheet_name='GFM')

df_routing_GFC.rename(columns={
    'Crew\nTotal \nRevised':'Crew\nTotal '
},inplace=True)

# Merging all datasets into one combined DataFrame.
df_routing=pd.concat([df_routing_NFC1, df_routing_NFC2, df_routing_GFC, df_routing_KFC], ignore_index=True)

df_routing = df_routing[['Unit', 'Item Code', 'Crew\nTotal ', 'Run Rate (Primary UOM)\nFG', 'Power Usage\nTotal ']].rename(columns={
    'Crew\nTotal ': 'Crew Total',
    'Run Rate (Primary UOM)\nFG': 'Run Rate (Primary UOM FG)',
    'Power Usage\nTotal ': 'Power Usage FG'
})

df_production

# Define Exchange Rates

# default_exchange_rates = {'KFC': 3.26796317659, 'NFC': 0.26666666666, 'GFC': 0.2722940776}
## Applying updated Exchange rates for year 2023
default_exchange_rates = {'KFC': 3.25520833333, 'NFC': 0.26666666666, 'GFC': 0.2722940776}

df_routing.columns

# Preprocessing Steps

df_routing.head()

df_routing_missing.head()

# Preprocessing Steps for Production File
# Filter the DataFrame based on the 'retain_flag' column
df_production = df_production[df_production['retain_flag'] == 1]

# Select specific columns
selected_columns = ['FGSKUCode', 'Description', 'Factory', 'Line', 'Throughput', 'retain_flag']
df_production = df_production[selected_columns]

df_production.head()

df_routing.head()

# Create the FACTCode column based on conditions and choices
conditions = [
    df_routing['Unit'].isin(['NF1', 'NF2']),
    df_routing['Unit'] == 'GFM',
    df_routing['Unit'] == 'KPF'
]
choices = ['NFC', 'GFC', 'KFC']
df_routing['FACTCode'] = np.select(conditions, choices, default=None)
######################## Add missing data#############################

df_routing=pd.concat([df_routing,df_routing_missing])
############################################################

# Preprocessing Steps for 3 Entries

# Rename columns
df_3.rename(columns={
    'Unnamed: 2': 'Type',
    'Unnamed: 3': 'Unit',
    'NFC': 'NFC(SAR)',
    'KFC': 'KFC(KD)',
    'GFC': 'GFC(AED)'
}, inplace=True)

# Filter data for 'Electricity' type
df_Electricity = df_3[df_3['Type'] == 'Electricity ']

# Melt DataFrame
df_Electricity = df_Electricity.melt(id_vars=['Type', 'Unit'], 
                                     value_vars=['NFC(SAR)', 'KFC(KD)', 'GFC(AED)'],
                                     var_name='Factory', 
                                     value_name='Cost')

# Split 'Factory' column
df_Electricity[['FACTCode', 'Currency']] = df_Electricity['Factory'].str.split('(', expand=True)

# Remove ')' from 'Currency' column
df_Electricity['Currency'] = df_Electricity['Currency'].str.rstrip(')')

# Select specific columns
df_Electricity = df_Electricity[['Type', 'FACTCode', 'Currency', 'Cost']]

# Calculate 'Cost (in USD)'
df_Electricity['Cost (in USD)'] = df_Electricity.apply(lambda row: row['Cost'] * default_exchange_rates.get(row['FACTCode'], 1), axis=1)

# Merge the Production and Routing files using the Fact-FG level

#df=pd.merge(df_production,df_routing,left_on=['Factory','FGSKUCode'],right_on=['FACTCode','Item Code'],how='left')
#df.head()
#[df['FGSKUCode']==4001971947]
df = pd.merge(df_production, df_routing, left_on=['Factory', 'FGSKUCode'], right_on=['FACTCode', 'Item Code'], how='left')
df = pd.merge(df, df_Electricity[['FACTCode', 'Cost (in USD)']], on='FACTCode', how='left')
df.head()

# Check duplicate 
def remove_duplicates_keep_first_occurrence(df):
    """
    Remove duplicates in the DataFrame based on 'FactCode', 'FGSKUCode', and 'Line',
    and keep only the first occurrence.
    
    Parameters:
    - df: Input DataFrame
    
    Returns:
    - DataFrame without duplicates based on specified columns
    """
    df_filtered = df.drop_duplicates(subset=['Factory', 'FGSKUCode', 'Line'], keep='first')
    
    return df_filtered

df=remove_duplicates_keep_first_occurrence(df)

df.head(1)

df=df[['Factory','FGSKUCode','Description','Line','Throughput','Crew Total','Run Rate (Primary UOM FG)','Power Usage FG','Cost (in USD)']]

df['EnergyCostPerUnit'] = (df['Power Usage FG'] * df['Cost (in USD)']) / df['Throughput']
df['MOQ'] = None

df = df[['Factory', 'Line', 'FGSKUCode', 'Throughput', 'EnergyCostPerUnit', 'Crew Total']]
df = df.rename(columns={
    'Factory': 'FactCode',
    'Line': 'LineNum',
    'FGSKUCode': 'Unified code',
    'Throughput': 'ThruputPerHour',
    'Crew Total': 'CrewSize'
})

# Add MOQ column (as it was in the original code)
df['MOQ'] = None

# Reorder columns to match the original output
df = df[['FactCode', 'LineNum', 'Unified code', 'MOQ', 'EnergyCostPerUnit', 'CrewSize', 'ThruputPerHour']]
### Adding Retain Flag Logic
df['retain_flag'] = df.apply(lambda row: 0 if pd.isnull(row['EnergyCostPerUnit']) or pd.isnull(row['CrewSize'])or pd.isnull(row['ThruputPerHour']) else 1, axis=1)



df.to_csv('df_final.csv')

# QC


df_final = df.copy()
#For QC Purpose
df_prod_uniquefg = df_final[['Unified code']].drop_duplicates()

# Print the unique DataFrame

df_merge= pd.merge(df_demand,df_prod_uniquefg, left_on='FGSKUCode',right_on='Unified code' ,how='inner')
df_merge

# Get the number of distinct values for each column
distinct_counts = df_final.nunique()

# Get the number of null values for each column
null_counts = df_final.isnull().sum()
counts=df_final.count()
# Get the count, mean, std, min, and percentile distribution for each column
summary_stats = df_final.describe(percentiles=[0.25, 0.5, 0.75]).transpose()

# Combine all the results into a single DataFrame
result_df = pd.DataFrame({
    'Count':counts,
    'Distinct Values': distinct_counts,
    'Null Count': null_counts,
    'Mean': summary_stats['mean'],
    'Std': summary_stats['std'],
    'Minimum Value': summary_stats['min'],
    'Percentile Distribution (25%)': summary_stats['25%'],
    'Percentile Distribution (50%)': summary_stats['50%'],
    'Percentile Distribution (75%)': summary_stats['75%'],
    'Maximum Value': summary_stats['max']
})



result_df

# FGSKU Count

df_month_demand_ret_1 = df_demand[df_demand['Ret_Flag'] == 1]
df_month_demand_ret_1=df_month_demand_ret_1['FGSKUCode'].nunique()

df_month_demand_merge_ret_1 = df_merge[df_merge['Ret_Flag'] == 1]
df_month_demand_merge_ret_1=df_month_demand_merge_ret_1['FGSKUCode'].nunique()


# VOLUME(CARTONS)


df_demand_carton=df_demand[df_demand['UM'] == 'CARTON']
df_demand_carton_merge =df_merge[df_merge['UM'] == 'CARTON']

df_demand_carton_ret_1 = df_demand_carton[df_demand_carton['Ret_Flag'] == 1]
df_demand_carton_ret_1=df_demand_carton_ret_1['Demand'].sum().sum()

df_demand_carton_merge_ret_1 = df_demand_carton_merge[df_demand_carton_merge['Ret_Flag'] == 1]
df_demand_carton_merge_ret_1 = df_demand_carton_merge_ret_1['Demand'].sum().sum()


# VOLUME(KG)


df_demand_kg=df_demand[df_demand['UM'] == 'KILO GRAM']
df_demand_merge_kg=df_merge[df_merge['UM'] == 'KILO GRAM']

#main
df_demand_kg_ret_1 = df_demand_kg[df_demand_kg['Ret_Flag'] == 1]
df_demand_kg_ret_1=df_demand_kg_ret_1['Demand'].sum().sum()

df_demand_merge_kg_ret_1 = df_demand_merge_kg[df_demand_merge_kg['Ret_Flag'] == 1]
df_demand_merge_kg_ret_1=df_demand_merge_kg_ret_1['Demand'].sum().sum()

#final table
#Storing result in a single df
result_df_1 = pd.DataFrame({
    'Demand Output (Ret_Flag=1)': [df_month_demand_ret_1,df_demand_carton_ret_1,df_demand_kg_ret_1]
   , 'Prod output': [df_month_demand_merge_ret_1,df_demand_carton_merge_ret_1,df_demand_merge_kg_ret_1]
}, index=['FGSKU','Volume (cartons)','Volume (kgs)'])

# Display the result 
print(result_df_1.to_string(float_format='{:,.2f}'.format))

dfs = [df_final, result_df, result_df_1]

# Create a new Excel workbook

wb = Workbook()
 
# Remove default sheet created by openpyxl

default_sheet = wb['Sheet']

wb.remove(default_sheet)

 
# Create a named style for bold headings

bold_style = NamedStyle(name='bold_style')

bold_style.font = Font(bold=True)  # Set font color to white
 
# Define the desired sheet names

sheet_names = ['Production_output', 'QC_output']
 
# Define blue fill

blue_fill = PatternFill(start_color="ADD8E6", end_color="ADD8E6", fill_type="solid")
 
# Loop through each DataFrame and add it to a new sheet in the workbook

for df, sheet_name in zip(dfs, sheet_names):

    # Create a new sheet

    ws = wb.create_sheet(title=sheet_name)
 
    # Write DataFrame to the sheet, starting from cell A1
   # Write header row

    if sheet_name == 'Production_output':

        header_row = [str(col) for col in df.columns]

        ws.append(header_row)

        for idx, row in df.iterrows():

            ws.append(row.tolist())

        for cell in ws[1]:

            cell.style = bold_style

            cell.fill = blue_fill

        for row in ws.iter_rows(min_row=1, max_row=ws.max_row, max_col=ws.max_column):

            for cell in row:

                cell.border = Border(left=Side(style='thin'),

                                     right=Side(style='thin'),

                                     top=Side(style='thin'),

                                     bottom=Side(style='thin'))

        ws.sheet_view.showGridLines = False

        for col in ws.columns:

            max_length = 0

            for cell in col:

                try:

                    if len(str(cell.value)) > max_length:

                        max_length = len(cell.value)

                except:

                    pass

            adjusted_width = (max_length + 2) * 1.2

            ws.column_dimensions[col[0].column_letter].width = adjusted_width
            # Add one row and column at the top
        ws.insert_rows(idx=1)
        ws.insert_cols(idx=1)
        ws.column_dimensions['A'].width = 3
 

    elif sheet_name == 'QC_output':

        df = result_df

        header_row = ['Column Name'] + [str(col) for col in df.columns]

        ws.append(header_row)

        for idx, row in df.iterrows():

            ws.append([idx]+row.tolist())

        for cell in ws[1]:

            cell.style = bold_style

            cell.fill = blue_fill

        for row in ws.iter_rows(min_row=1, max_row=ws.max_row, max_col=ws.max_column):

            for cell in row:

                cell.border = Border(left=Side(style='thin'),

                                     right=Side(style='thin'),

                                     top=Side(style='thin'),

                                     bottom=Side(style='thin'))

        for col in ws.columns:

            max_length = 0

            for cell in col:

                try:

                    if len(str(cell.value)) > max_length:

                        max_length = len(cell.value)

                except:

                    pass

            adjusted_width = (max_length + 2) * 1.2

            ws.column_dimensions[col[0].column_letter].width = adjusted_width
 
        for _ in range(4):

            ws.append([])

        df = result_df_1

        header_row = ['Data'] + [str(col) for col in df.columns]

        ws.append(header_row)

        for idx, row in df.iterrows():

            ws.append([idx]+row.tolist())

        for idx, cell in enumerate(ws[14], start=1):

                if idx <= 3:

                    cell.style = bold_style

                    cell.fill = blue_fill

        for row in ws.iter_rows(min_row=14, max_row=17, max_col=3):

            for cell in row:

                cell.border = Border(left=Side(style='thin'),

                                     right=Side(style='thin'),

                                     top=Side(style='thin'),

                                     bottom=Side(style='thin'))

        ws.sheet_view.showGridLines = False

        for col in ws.columns:

            max_length = 0

            for cell in col:

                try:

                    if len(str(cell.value)) > max_length:

                        max_length = len(cell.value)

                except:

                    pass

            adjusted_width = (max_length + 2) * 1.2

            ws.column_dimensions[col[0].column_letter].width = adjusted_width
# Add one row and column at the top
ws.insert_rows(idx=1)
ws.insert_cols(idx=1)
ws.column_dimensions['A'].width = 3
# Save the workbook

wb.save('03-05-2024_Prod_Automation.xlsx')
print("Completed Successfully")
