==> Script08_FactoryLabourPrep.ipynb <==
# Importing all the dependencies
import pandas as pd
import itertools
import os

from openpyxl import load_workbook
from openpyxl.styles import PatternFill, Border, Side, Font
from openpyxl.utils import get_column_letter
from openpyxl.worksheet.table import Table, TableStyleInfo
from openpyxl.styles.alignment import Alignment

###Input Files
##This is the configuraiton and the only thing to change in this code whenever you rerun
cwd = os.getcwd()
standard_costs =os.path.join(cwd, "client_data", "Data 3 entities.xlsx")

##### Defining default exchange rates #############
# def_er= {'KFC': 3.26796317659, 'NFC': 0.26666666666, 'GFC': 0.2722940776}

## Update with 2023 Exchange rates 

def_er= {'KFC': 3.25520833333, 'NFC': 0.26666666666, 'GFC': 0.2722940776}
# Convert the dictionary to a DataFrame
exchange_rates = pd.DataFrame(list(def_er.items()), columns=['Factory', 'Value'])
# exchange_rates

#### Import input data - Data 3 entities
df_3=pd.read_excel(standard_costs,skiprows=2)

##### Preprocessing Steps for 3 Entries

# Rename columns
df_3.rename(columns={
    'Unnamed: 2': 'Type',
    'Unnamed: 3': 'Unit',
    'NFC': 'NFC(SAR)',
    'KFC': 'KFC(KD)',
    'GFC': 'GFC(AED)'
}, inplace=True)
# Remove extra blank columns
df_3 = df_3.dropna(axis=1, how='all')

# Filter data for multiple types
types_to_filter = ["Production Labour Cost FTE (CTC) - No OT", "Outsource Labour Cost ","Overtime ","Fixed Head Count (DVL) -2023"] 
df_filtered = df_3[df_3['Type'].isin(types_to_filter)]
# df_3['Type'].unique
# df_filtered

# Melt DataFrame
df_altered = df_filtered.melt(id_vars=['Type', 'Unit'], 
                                     value_vars=['NFC(SAR)', 'KFC(KD)', 'GFC(AED)'],
                                     var_name='Factory', 
                                     value_name='Cost')
# Split 'Factory' column
df_altered[['FACTCode', 'Currency']] = df_altered['Factory'].str.split('(', expand=True)

# Remove ')' from the 'Currency' column
df_altered['Currency'] = df_altered['Currency'].str.rstrip(')')
# Select specific columns
df_altered = df_altered[['Type', 'FACTCode', 'Currency', 'Cost']]
df_altered = pd.merge(df_altered, exchange_rates, left_on=['FACTCode'], 
                                   right_on=['Factory'], how='left')
df_altered['Cost (in USD)']= df_altered['Cost']*df_altered['Value']

# Group the DataFrame by the 'Type' column
grouped = df_altered.groupby('Type')

# Initialize an empty dictionary to store the DataFrames
grouped_dfs = {}

# Iterate over each group and store it in the dictionary
for type_name, group in grouped:
    grouped_dfs[type_name] = group.copy()  # Make a copy to ensure the independence of the original DataFrame

grouped_dfs

hc_df = grouped_dfs['Fixed Head Count (DVL) -2023']
ot_df = grouped_dfs['Outsource Labour Cost ']
ot_reg_df = grouped_dfs['Overtime ']
reg_df = grouped_dfs['Production Labour Cost FTE (CTC) - No OT']
# reg_df

##### Outsource Labour Cost and convert that for hr ##########
# Create a dictionary with the provided data
data = {
    'NFC': [3700],
    'KFC': [250]
}

# Create the DataFrame
df_OT = pd.DataFrame(data, index=['Outsource Labour Cost'])

# Transpose the DataFrame
df_OT = df_OT.T

# Add additional columns
df_OT['Type'] = 'Outsource Labour Cost'
df_OT['Unit'] = 'per month'

# Reorder the columns
df_OT = df_OT[['Type', 'Unit', 'Outsource Labour Cost']]
df_OT['Outsource Labour Cost -per hr'] =df_OT['Outsource Labour Cost']/(24*8)
ot_df



df_OT['Factory'] = df_OT.index
# Define a function to apply the conditions
def implement_value(Factory):
    if Factory == 'NFC':
        return 0.26666666666
    elif Factory == 'KFC':
        return 3.26796317659
    else:
        return None

# Apply the function to create the 'implement' column
df_OT['implement'] = df_OT['Factory'].apply(implement_value)
df_OT['Cost (in USD)']= df_OT['implement']*df_OT['Outsource Labour Cost -per hr']
# df_OT.rename(columns={ 'Outsource Labour Cost -per hr':'R_OTWages'})
df_OT

# Concatenating DataFrames
ot_c_df = pd.concat([ot_df[['Factory', 'Cost (in USD)']], df_OT[['Factory', 'Cost (in USD)']]])
# Drop rows where cost is NaN
ot_c_df = ot_c_df.dropna(subset=['Cost (in USD)']).reset_index()
ot_c_df.columns

####### Prepare the base data ###########################

# List of factories and months
factories = ['NFC', 'KFC', 'GFC']
months = list(range(1, 13))  # Months from 1 to 12

# Generate all combinations
combinations = list(itertools.product(factories, months))

# Convert to DataFrame - This is the base data
df = pd.DataFrame(combinations, columns=['Factory', 'Month'])


The following assumptions has been considered while preparing the labour data 

**Assumptions:**

1. 23 working days/month
2. 8 Regular hours and 2 hours of Overtime (Regular labor)
3. Minimum Regular labor - 80% of labor count (-20%)
4. Maximum Regular labour - 120% of labor count (+20%)
5. 0 Regular hours and 8 hours of Over time(Contract labor) 
6. Contract Labor count - 20 
7. Min Contract labors - 0 
8. Max Contract labors - 50

#### Adding Regular labour Count, Labour Days, Labour Hours and Overtime hours
df_base = pd.merge(df, hc_df[['FACTCode','Cost']], left_on=['Factory'], 
                                   right_on=['FACTCode'], how='left')
df_base = df_base[['Factory', 'Month','Cost']]
df_base = df_base.rename(columns={'Cost':'R_Count'})

# Calculate the minimum and maximum regular labour based on the 'R_Count' column
df_base['R_Min'] = (0.8 * df_base['R_Count']).apply(round)  # 80% of labor count
df_base['R_Max'] = (1.2 * df_base['R_Count']).apply(round)  # 120% of labor count

#### Map Wages #####

df_base = pd.merge(df_base, ot_reg_df[['FACTCode','Cost (in USD)']], left_on=['Factory'], 
                                   right_on=['FACTCode'], how='left')

df_base = df_base[['Factory', 'Month', 'R_Count', 'R_Min', 'R_Max','Cost (in USD)']]
df_base = df_base.rename(columns={ 'Cost (in USD)':'R_OTWages'})

#### Map regular wages

df_base = pd.merge(df_base, reg_df[['FACTCode','Cost (in USD)']], left_on=['Factory'], 
                                   right_on=['FACTCode'], how='left')

df_base = df_base[['Factory', 'Month', 'R_Count', 'R_Min', 'R_Max','R_OTWages','Cost (in USD)']]
df_base = df_base.rename(columns={ 'Cost (in USD)':'R_Wages'})

#### Adding Fixed value columns####
df_base['R_Days'] =24
df_base['R_Hrs'] = 8
df_base['R_OT'] = 4
df_base['C_Days'] =24
df_base['C_Hrs'] = 0
df_base['C_OT'] = 8
df_base['C_Count']=20
df_base['C_Min'] = 0
df_base['C_Max'] = 50
df_base['C_Wages'] = 0

## Map C_OTWages 
df_base = pd.merge(df_base, ot_c_df[['Factory', 'Cost (in USD)']], left_on=['Factory'], 
                                   right_on=['Factory'], how='left')

df_base = df_base.rename(columns={ 'Cost (in USD)':'C_OTWages','Factory':'FactCode','Month':'MonthNum'})
df_base = df_base[['FactCode','MonthNum','R_Days','R_Count','R_Min','R_Max','R_Hrs','R_OT','R_Wages','R_OTWages','C_Days','C_Count','C_Min','C_Max','C_Hrs','C_OT','C_Wages','C_OTWages']]


####### Formatting Code #####

# Write the name of dataframe and tab name that you want in excel
dataframes = {
    'Labour Data': df_base,
   }

# Specify the path to save your Excel file
file_path = r'22-04-2024 labour Data.xlsx'
# Using ExcelWriter to write each DataFrame to a separate sheet
with pd.ExcelWriter(file_path, engine='openpyxl') as writer:
    for sheet_name, dataframe in dataframes.items():
        dataframe.to_excel(writer, sheet_name=sheet_name, index=False)

    # Load the workbook and apply formatting
    workbook = writer.book
    for sheet_name in dataframes.keys():
        worksheet = workbook[sheet_name]
        worksheet.sheet_view.showGridLines = False  # Remove gridlines

        # Add one row and column at the top
        worksheet.insert_rows(1)
        worksheet.insert_cols(1)
        worksheet.cell(row=1, column=1)

        # Add table borders
        border = Border(left=Side(style='thin'), 
                        right=Side(style='thin'), 
                        top=Side(style='thin'), 
                        bottom=Side(style='thin'))
        for row in worksheet.iter_rows(min_row=2, min_col=2, max_col=worksheet.max_column):
            for cell in row:
                cell.border = border

        
# Set light blue color for headers
        for cell in worksheet.iter_cols(min_row=2, min_col=2):
            cell[0].fill = PatternFill(start_color="ADD8E6", end_color="ADD8E6", fill_type="solid")  # Light blue
    
        # Remove borders for the extra added row and column at the top
        for cell in worksheet['A1:C1'][0]:
            cell.border = None
# Autofit columns
        for col in worksheet.columns:
            max_length = 0
            column = col[0].column_letter  # Get the column name
            for cell in col:
                try:  # Necessary to avoid error on empty cells
                    if len(str(cell.value)) > max_length:
                        max_length = len(cell.value)
                except:
                    pass
            adjusted_width = (max_length + 2) * 1.2
            worksheet.column_dimensions[column].width = adjusted_width
 # Convert pixels to character units
        

        # Apply auto filters starting from the 2nd row in the B column
        #worksheet.auto_filter.ref = worksheet.dimensions
        last_row = worksheet.max_row
        last_column = worksheet.max_column
        range_str = f"B2:{get_column_letter(last_column)}{last_row}"
        worksheet.auto_filter.ref = range_str
# Save the Excel file
workbook.save(file_path)

print("All DataFrames are written to {}".format(file_path))
