==> Script09_SKUMasterPrep.ipynb <==
### Import Dependencies
import pandas as pd
import os
import numpy as np
from openpyxl import load_workbook
from openpyxl.styles import PatternFill, Border, Side, Font
from openpyxl.utils import get_column_letter
from openpyxl.worksheet.table import Table, TableStyleInfo
from openpyxl.styles.alignment import Alignment

cwd = os.getcwd()
demand_filepath = "28-04-2024_Demand_Automation.xlsx"

NFC_master_file = os.path.join(cwd, "client_data", "AMF_Inventory_Item_Master___Al_050224.xlsx")
KFC_master_file = os.path.join(cwd, "client_data", "Master file on 11-02-2024.xlsx")
GFC_master_file = os.path.join(cwd, "client_data", "AMF_Inventory_Item_Master___Al_080224.xlsx")
budget_file = os.path.join(cwd, "client_data", "PROTEIN_COGS_FORECAST-Q4+2024_BUDGET-TP3_V16.01.xlsx")

trimming_output = "29-04-24_Trimming_Output.csv"

# import pandas as pd

# df_bom = pd.read_excel("BOM Top 80 Output.xlsx", skiprows=1)  # Row 5 in Excel


# df_bom.head()


#####RMSKU Master List Input
df_bom= pd.read_excel(r"BOM Top 80 Output.xlsx",skiprows=1)
df_bom = df_bom.drop(df_bom.columns[0], axis=1)
df_rm= pd.read_excel(r"RM Supply.xlsx",skiprows=1)
df_rm = df_rm.drop(df_rm.columns[0], axis=1)

#Taking all unique combination of RMCode , FactCode and RM UOM from RM Supply
df_rm_unique = df_rm[['RM Code', 'RM UOM', 'FactCode']].drop_duplicates()
#Taking all unique combination of RMCode and RM Description UOM from Top 80% BOM
df_bom_unique= df_bom[['RM CODE','RM DESC']].drop_duplicates()
df_bom_unique.columns= ['RM Code','RM Desc']

#Merging both datasets
df_merge= pd.merge( df_rm_unique,df_bom_unique, on='RM Code', how= 'left')

dataframes = {
    'RMSKU Master': df_merge,
   }

# Specify the path to save your Excel file
file_path = r'03-05-2024 RMSKU Master List.xlsx'
# Using ExcelWriter to write each DataFrame to a separate sheet
with pd.ExcelWriter(file_path, engine='openpyxl') as writer:
    for sheet_name, dataframe in dataframes.items():
        dataframe.to_excel(writer, sheet_name=sheet_name, index=False)

    # Load the workbook and apply formatting
    workbook = writer.book
    for sheet_name in dataframes.keys():
        worksheet = workbook[sheet_name]
        worksheet.sheet_view.showGridLines = False  # Remove gridlines

        # Add one row and column at the top
        worksheet.insert_rows(1)
        worksheet.insert_cols(1)
        worksheet.cell(row=1, column=1)

        # Add table borders
        border = Border(left=Side(style='thin'), 
                        right=Side(style='thin'), 
                        top=Side(style='thin'), 
                        bottom=Side(style='thin'))
        for row in worksheet.iter_rows(min_row=2, min_col=2, max_col=worksheet.max_column):
            for cell in row:
                cell.border = border

        
# Set light blue color for headers
        for cell in worksheet.iter_cols(min_row=2, min_col=2):
            cell[0].fill = PatternFill(start_color="ADD8E6", end_color="ADD8E6", fill_type="solid")  # Light blue
    
        # Remove borders for the extra added row and column at the top
        for cell in worksheet['A1:C1'][0]:
            cell.border = None
# Autofit columns
        for col in worksheet.columns:
            max_length = 0
            column = col[0].column_letter  # Get the column name
            for cell in col:
                try:  # Necessary to avoid error on empty cells
                    if len(str(cell.value)) > max_length:
                        max_length = len(cell.value)
                except:
                    pass
            adjusted_width = (max_length + 2) * 1.2
            worksheet.column_dimensions[column].width = adjusted_width
 # Convert pixels to character units
        

        # Apply auto filters starting from the 2nd row in the B column
        #worksheet.auto_filter.ref = worksheet.dimensions
        last_row = worksheet.max_row
        last_column = worksheet.max_column
        range_str = f"B2:{get_column_letter(last_column)}{last_row}"
        worksheet.auto_filter.ref = range_str
# Save the Excel file
workbook.save(file_path)

print("All DataFrames are written to {}".format(file_path))

######################### Importing Input Files ##########################
### Demand file to get the list of FGs and UOM
Demand_file = pd.read_excel(demand_filepath)

###Source of WtPerUnit
NFC_master = pd.read_excel(NFC_master_file)
KFC_master = pd.read_excel(KFC_master_file)
GFC_master = pd.read_excel(GFC_master_file)

########## Source For Booking Cost 
Booking_cost = pd.read_excel(budget_file, sheet_name='FG',usecols="A:F,IW:KO", skiprows = 4)

###### Source for trimming generated per KG ###############################

Trim_prod =  pd.read_csv(trimming_output)
# Trim_prod.columns

####### Logic: If trimming is being produced in multiple factories in multiple level, in that case we will take average values ###############
# Create the pivot table
trimming_table = Trim_prod.pivot_table(index='ASSEMB_ITEM_CODE', values='Trimming(in KG)', aggfunc='mean')
# trimming_table.head()

### The FG base list will be derived from Demand where retain flag is 1
Demand_Base = Demand_file[Demand_file['Ret_Flag']==1]
FG_List = Demand_Base[['FGSKUCode','UM']].drop_duplicates()

##### Map SUb Category, Description, Wt UOM from combined master
master = pd.concat([NFC_master[[' Item Code',' Description','Weight UOM','OPM Sub Category']], KFC_master[[' Item Code',' Description','Weight UOM','OPM Sub Category']],GFC_master[[' Item Code',' Description','Weight UOM','OPM Sub Category']]])
# master.head()
FG_Master = pd.merge(FG_List, master, left_on=['FGSKUCode'], 
                                   right_on=[' Item Code'], how='left')

### Average Booking cost is taken for all the months and accross all the factories and then mapped 

Bk_Cst_base =Booking_cost[['Entity','Item Code','Description','UOM','Currency','OH','Total Cost.60','Total Cost.61','Total Cost.62','Total Cost.63','Total Cost.64','Total Cost.65','Total Cost.66','Total Cost.67','Total Cost.68','Total Cost.69','Total Cost.70','Total Cost.71']]
Bk_Cst_base['Avg_Budget_cost']=Bk_Cst_base[['Total Cost.60','Total Cost.61','Total Cost.62','Total Cost.63','Total Cost.64','Total Cost.65','Total Cost.66','Total Cost.67','Total Cost.68','Total Cost.69','Total Cost.70','Total Cost.71']].fillna(0).astype('float64').mean(axis= 1)
Bk_Cst_base = Bk_Cst_base[['Entity', 'Item Code', 'Description', 'UOM', 'Currency','Avg_Budget_cost']]
Bk_Cst_base1 = Bk_Cst_base.groupby(['Item Code','UOM', 'Currency'])['Avg_Budget_cost'].mean()

###UOMPerPallet and trimmingKGperunit kept blank as logic is not clear yet
FG_Master_base = pd.merge(FG_Master, Bk_Cst_base1, left_on=['FGSKUCode'], 
                                   right_on=['Item Code'], how='left')
FG_Master_base =FG_Master_base[['FGSKUCode', ' Description','OPM Sub Category','UM','Weight UOM','Avg_Budget_cost']].drop_duplicates()

# Merge the pivot table with FG_master_base based on ASSEMB_ITEM_CODE and FGSKUCode
FG_Master_base = pd.merge(trimming_table, FG_Master_base, left_index=True, right_on='FGSKUCode', how='right')
FG_Master_base['UOMPerPallet'] =np.nan 

FG_Master_base =FG_Master_base[['FGSKUCode', ' Description','OPM Sub Category','UM','Weight UOM','UOMPerPallet','Avg_Budget_cost','Trimming(in KG)']]
# Rename columns
FG_Master_base = FG_Master_base.rename(columns={'FGSKUCode':'FG Code', 
' Description':'FG Desc', 
'OPM Sub Category':'FG Cat', 
'UM':'FG UOM', 
'Weight UOM':'WtPerUnit',
'UOMPerPallet':'UOMPerPallet', 
'Avg_Budget_cost':'Book Cost', 
'Trimming(in KG)':'TrimKGPerUnit'})
# FG_Master_base['retain_flag'] = FG_Master_base.apply(lambda row: 0 if pd.isnull(row['WtPerUnit']) or pd.isnull(row['Book Cost']) else 1, axis=1)
# FG_Master_base
FG_Master_base['retain_flag']=1



###################### Formatting and final Output ####################################

# Write the name of dataframe and tab name that you want in excel
dataframes = {
    'FG SKU': FG_Master_base,
   }

# Specify the path to save your Excel file
file_path = r'29_04_2024_FG-SKU.xlsx'
# Using ExcelWriter to write each DataFrame to a separate sheet
with pd.ExcelWriter(file_path, engine='openpyxl') as writer:
    for sheet_name, dataframe in dataframes.items():
        dataframe.to_excel(writer, sheet_name=sheet_name, index=False)

    # Load the workbook and apply formatting
    workbook = writer.book
    for sheet_name in dataframes.keys():
        worksheet = workbook[sheet_name]
        worksheet.sheet_view.showGridLines = False  # Remove gridlines

        # Add one row and column at the top
        worksheet.insert_rows(1)
        worksheet.insert_cols(1)
        worksheet.cell(row=1, column=1)

        # Add table borders
        border = Border(left=Side(style='thin'), 
                        right=Side(style='thin'), 
                        top=Side(style='thin'), 
                        bottom=Side(style='thin'))
        for row in worksheet.iter_rows(min_row=2, min_col=2, max_col=worksheet.max_column):
            for cell in row:
                cell.border = border

        
# Set light blue color for headers
        for cell in worksheet.iter_cols(min_row=2, min_col=2):
            cell[0].fill = PatternFill(start_color="ADD8E6", end_color="ADD8E6", fill_type="solid")  # Light blue
    
        # Remove borders for the extra added row and column at the top
        for cell in worksheet['A1:C1'][0]:
            cell.border = None
# Autofit columns
        for col in worksheet.columns:
            max_length = 0
            column = col[0].column_letter  # Get the column name
            for cell in col:
                try:  # Necessary to avoid error on empty cells
                    if len(str(cell.value)) > max_length:
                        max_length = len(cell.value)
                except:
                    pass
            adjusted_width = (max_length + 2) * 1.2
            worksheet.column_dimensions[column].width = adjusted_width
 # Convert pixels to character units
        

        # Apply auto filters starting from the 2nd row in the B column
        #worksheet.auto_filter.ref = worksheet.dimensions
        last_row = worksheet.max_row
        last_column = worksheet.max_column
        range_str = f"B2:{get_column_letter(last_column)}{last_row}"
        worksheet.auto_filter.ref = range_str
# Save the Excel file
workbook.save(file_path)

print("All DataFrames are written to {}".format(file_path))
