==> Script14_UpdateDatabaseAfterQC.ipynb <==
# import pyodbc
# print("✅ pyodbc is installed")


# from sqlalchemy import create_engine, NVARCHAR, INTEGER, FLOAT
# print("✅ sqlalchemy is installed")


# !pip show pandas pyodbc sqlalchemy


import pandas as pd
import pyodbc
from sqlalchemy import create_engine
from sqlalchemy import create_engine, NVARCHAR, INTEGER, FLOAT


##Configuration of final output
combined_file_name = "09-05-2024 Full Data_v0.9.xlsx"

# List all available ODBC drivers
drivers = pyodbc.drivers()
print("Available ODBC drivers:")
for driver in drivers:
    print(driver)

# import subprocess

# def list_sql_servers():
#     result = subprocess.run(["sqlcmd", "-L"], capture_output=True, text=True)
#     print(result.stdout)

# list_sql_servers()

# database = '09052024_AmericanaFoods_BacktestingData_v0.9'
# # 

# Set up the connection parameters
server = 'DESKTOP-JNFPIJ4\\SQLEXPRESS'
database = '09052024_AmericanaFoods_BacktestingData_v0.9'
driver = 'ODBC Driver 17 for SQL Server'

# Build the connection string with Windows authentication
connection_string = f"DRIVER={driver};SERVER={server};DATABASE={database};trusted_connection=yes"

# Connect to the SQL Server database
conn = pyodbc.connect(connection_string)

# Query the database to get all table names
sql_query = "SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE = 'BASE TABLE'"
table_names_df = pd.read_sql(sql_query, conn)
#sql_query = "SELECT * FROM TC_06_NegativeList_01_FGSupply"
#df = pd.read_sql(sql_query, conn)
# Close the connection
conn.close()

# Display the table names DataFrame
print(table_names_df)
#print(df.head())

##Testing

### Making Changes from df to SQL Server

####  Step 1.a: Uploading All excel files¶

# Use raw string with 'r' prefix
excel_file = combined_file_name

# Read Excel file into DataFrames
df_lines = pd.read_excel(excel_file, sheet_name='Lines', skiprows=1)
df_country = pd.read_excel(excel_file, sheet_name='Country', skiprows=1)
df_demand = pd.read_excel(excel_file, sheet_name='Demand', skiprows=1)
df_Secondary_Freight = pd.read_excel(excel_file, sheet_name='Secondary Freight', skiprows=1)
df_warehouse = pd.read_excel(excel_file, sheet_name='Warehouse', skiprows=1)
df_fgsku = pd.read_excel(excel_file, sheet_name='FG SKUs', skiprows=1)
df_rmsku = pd.read_excel(excel_file, sheet_name='RM SKUs', skiprows=1)
df_trimming = pd.read_excel(excel_file, sheet_name='TrimmingSKUs', skiprows=1)
# df_WCC = pd.read_excel(excel_file, sheet_name='WorkingCaptitalCost', skiprows=1)
df_FG_M1OS = pd.read_excel(excel_file, sheet_name='FG - M1OS', skiprows=1)
df_RM_M1OS = pd.read_excel(excel_file, sheet_name='RM- M1OS', skiprows=1)
df_Primary_Freight = pd.read_excel(excel_file, sheet_name='Primary Freight', skiprows=1)
df_Production = pd.read_excel(excel_file, sheet_name='Production', skiprows=1)
df_BOM = pd.read_excel(excel_file, sheet_name='BOM', skiprows=1)
df_Factory_Operations = pd.read_excel(excel_file, sheet_name='Factory Operations', skiprows=1)
df_calender = pd.read_excel(excel_file, sheet_name='Calendar', skiprows=1)
df_RM_Supply = pd.read_excel(excel_file, sheet_name='RM Supply', skiprows=1)
# df_FG_Inventory_Norms = pd.read_excel(excel_file, sheet_name='FG Inventory Norms', skiprows=2)
df_LabourData = pd.read_excel(excel_file, sheet_name='LabourData', skiprows=1)
# df_FG_Restrictions = pd.read_excel(excel_file, sheet_name='FG Restrictions', skiprows=4)
# df_RM_Restrictions = pd.read_excel(excel_file, sheet_name='RM Restrictions', skiprows=4)
df_factory= pd.read_excel(excel_file, sheet_name='Factory', skiprows=1)
df_rm_transfer= pd.read_excel(excel_file, sheet_name='RM Transfer', skiprows=1)
df_fgfact_trimm=pd.read_excel(excel_file, sheet_name='FG- Fact Trimming', skiprows=1)
df_wh_fg=pd.read_excel(excel_file, sheet_name='WH-FG Storage', skiprows=1)
df_rmtransfer_restrictions=pd.read_excel(excel_file, sheet_name='RM Transfer Restrictions', skiprows=1)
df_min_contract_labour=pd.read_excel(excel_file, sheet_name='Minimum Contract Labour Hours', skiprows=1)
df_prep_line_fg=pd.read_excel(excel_file, sheet_name='Preparation Lines FG', skiprows=1)
df_prep_line_capacity=pd.read_excel(excel_file, sheet_name='Preparation Lines Capacity', skiprows=1)
df_TC_04_INV_06_FGCSInvNorm=pd.read_excel(excel_file,sheet_name='TC_04_INV_06_FGCSInvNorm',skiprows=1)
df_TC_04_INV_03_FGInv=pd.read_excel(excel_file,sheet_name='TC_04_INV_03_FGInv',skiprows=1)
df_TC_06_NegativeList_01_FGSupply=pd.read_excel(excel_file,sheet_name='TC_06_NegativeList_01_FGSupply',skiprows=1)
df_TC_06_NegativeList_02_RMSupply=pd.read_excel(excel_file,sheet_name='TC_06_NegativeList_02_RMSupply',skiprows=1)
df_TM_04_FIN_01_ICC=pd.read_excel(excel_file,sheet_name='TM_04_FIN_01_ICC',skiprows=1)
df_TM_03_CAL_01_Months=pd.read_excel(excel_file,sheet_name='TM_03_CAL_01_Months',skiprows=1)

# List of DataFrames
dataframes = [df_lines, df_country, df_demand, df_Secondary_Freight, df_warehouse, df_fgsku, df_rmsku, df_trimming, 
             df_FG_M1OS, df_RM_M1OS, df_Primary_Freight, df_Production, df_BOM, df_Factory_Operations, 
              df_calender, df_RM_Supply, df_LabourData, df_factory,df_rm_transfer,df_fgfact_trimm,df_wh_fg,df_rmtransfer_restrictions,
             df_min_contract_labour, df_prep_line_fg, df_prep_line_capacity,df_TC_04_INV_06_FGCSInvNorm ,df_TC_04_INV_03_FGInv,
             df_TC_06_NegativeList_01_FGSupply,df_TC_06_NegativeList_02_RMSupply,df_TM_04_FIN_01_ICC,df_TM_03_CAL_01_Months]


table_names = ['df_lines', 'df_country', 'df_demand', 'df_Secondary_Freight', 'df_warehouse', 'df_fgsku', 'df_rmsku', 'df_trimming', 
             'df_FG_M1OS', 'df_RM_M1OS', 'df_Primary_Freight', 'df_Production', 'df_BOM', 'df_Factory_Operations', 
               'df_calender', 'df_RM_Supply', 'df_LabourData','df_factory', 'df_rm_transfer','df_fgfact_trimm','df_wh_fg','df_rmtransfer_restrictions',
             'df_min_contract_labour', 'df_prep_line_fg', 'df_prep_line_capacity','df_TC_04_INV_06_FGCSInvNorm','df_TC_04_INV_03_FGInv',
              'df_TC_06_NegativeList_01_FGSupply','TC_06_NegativeList_02_RMSupply','df_TM_04_FIN_01_ICC','df_TM_03_CAL_01_Months']

# Remove the first unnamed column from each DataFrame
for df in dataframes:
    if 'Unnamed: 0' in df.columns:
        df.drop('Unnamed: 0', axis=1, inplace=True)

# Now you can print the table name followed by the DataFrame
for table_name, df in zip(table_names, dataframes):
    print(f"Table Name: {table_name}")
    print(df.head())
    print("\n")


# Remove the first unnamed column from each DataFrame
#for df in dataframes:
 #   df.drop(df.columns[0], axis=1, inplace=True)
#print(df_Primary_Freight.head())
# Now you can work with the DataFrames
#for df in dataframes:
 #   print(df.head())
    

#### Step 1.b: Making tables as per Jaideep's Tables in the Database

# df_demand
df_demand = df_demand[df_demand['Ret_Flag'] == 1]
df_demand.drop(columns=['Ret_Flag'], inplace=True)

# df_Secondary_Freight
df_Secondary_Freight = df_Secondary_Freight[df_Secondary_Freight['retain_flag'] == 1]
df_Secondary_Freight.drop(columns=['retain_flag','Missing_Flag'], inplace=True)

# df_Primary_Freight
df_Primary_Freight = df_Primary_Freight[df_Primary_Freight['retain_flag'] == 1]
df_Primary_Freight.drop(columns=['retain_flag','Missing_Flag'], inplace=True)

# df_BOM
df_BOM = df_BOM[df_BOM['retain_flag'] == 1]
df_BOM.drop(columns=['retain_flag'], inplace=True)

# df_FG_M1OS
df_FG_M1OS = df_FG_M1OS[df_FG_M1OS['retain_flag'] == 1]
df_FG_M1OS.drop(columns=['retain_flag','Missing_Flag'], inplace=True)
 
# df_RM_M1OS
df_RM_M1OS = df_RM_M1OS[df_RM_M1OS['Retain_Flag'] == 1]  
df_RM_M1OS.drop(columns=['Retain_Flag','Missing_Flag'], inplace=True)
 
# df_fgsku
df_fgsku = df_fgsku[df_fgsku['retain_flag'] == 1]
df_fgsku.drop(columns=['retain_flag'], inplace=True)

# df_RM_Supply
df_RM_Supply = df_RM_Supply[df_RM_Supply['retain_flag'] == 1]
df_RM_Supply.drop(columns=['retain_flag'], inplace=True)

# df_Production
df_Production = df_Production[df_Production['retain_flag'] == 1]
df_Production.drop(columns=['retain_flag'], inplace=True)   


# df_wh_fg
df_wh_fg = df_wh_fg[df_wh_fg['retain_flag'] == 1]
df_wh_fg.drop(columns=['retain_flag'], inplace=True)


####  Step 2: Storing all df in required database¶

# Set up the connection parameters
# Set up the connection parameters
server = 'DESKTOP-JNFPIJ4\\SQLEXPRESS'
database = '09052024_AmericanaFoods_BacktestingData_v0.9'
driver = 'ODBC Driver 17 for SQL Server'

# Build the connection string with Windows authentication
connection_string = f"mssql+pyodbc://{server}/{database}?driver={driver}"

# Create SQLAlchemy engine
engine = create_engine(connection_string)

# Define your DataFrames here:

dfs = {
    'TM_02_LOC_04_FactLine': df_lines,
    'TM_02_LOC_01_Country': df_country,
    'TC_01_Demand_01_FGDemand': df_demand,
    'TC_02_FRT_01_Sec': df_Secondary_Freight,
    'TC_02_FRT_02_Pri': df_Primary_Freight,
    'TC_03_Factories_01_BOM': df_BOM,
    'TC_03_Factories_02_WorkingDays': df_calender,
    'TC_03_Factories_03_ShiftData': df_Factory_Operations,
    'TC_03_Factories_04_LineFGSKUData': df_Production,
    'TC_03_Factories_06_LabourSupply': df_LabourData,
    'TC_04_INV_01_FGOS': df_FG_M1OS,
    'TC_04_INV_02_RMOS': df_RM_M1OS,
    'TC_04_INV_04_RMInv': df_RM_Supply,
    'TM_01_Material_01_FGSKU': df_fgsku,
    'TM_01_Material_02_RMSKU': df_rmsku,
    'TM_01_Material_03_TrimSKU': df_trimming,
    'TM_02_LOC_02_Warehouse': df_warehouse,
    'TM_02_LOC_03_Factory': df_factory,
    'T_01_RM_transfer': df_rm_transfer,
    'T_01_FGfact_trimm':df_fgfact_trimm,
    'T_01_WH_FG':df_wh_fg,
    'T_01_RMTransfer_Restrictions':df_rmtransfer_restrictions,
    'T_01_Min_contract_labour':df_min_contract_labour,
    'T_01_Prep_line_FG':df_prep_line_fg,
    'T_01_Prep_line_capacity':df_prep_line_capacity,
    'TC_04_INV_06_FGCSInvNorm':df_TC_04_INV_06_FGCSInvNorm,
    'TC_04_INV_03_FGInv':df_TC_04_INV_03_FGInv,
    'TC_06_NegativeList_01_FGSupply':df_TC_06_NegativeList_01_FGSupply,
    'TC_06_NegativeList_02_RMSupply':df_TC_06_NegativeList_02_RMSupply,
    'TM_04_FIN_01_ICC':df_TM_04_FIN_01_ICC,
    'TM_03_CAL_01_Months':df_TM_03_CAL_01_Months
    
}

tables_updated = []  # List to store the names of tables that were successfully updated

try:
    # Insert each DataFrame into its corresponding table
    for table_name, df in dfs.items():
        df.to_sql(table_name, con=engine, if_exists='replace', index=False)
        tables_updated.append(table_name)  # Add the table name to the list of updated tables
#If you want to create a new table with existing df that can be done by:
#df_WCC.to_sql('test_table', con=engine, if_exists='replace', index=False)

    # If no errors occurred, print all tables updated
    print("All tables updated successfully:")
    for table_name in tables_updated:
        print(table_name)

except Exception as e:
    print(f"An error occurred while inserting into table {table_name}: {e}")

finally:
    # Dispose the engine
    engine.dispose()


# Build the connection string with Windows authentication
connection_string = f"mssql+pyodbc://{server}/{database}?driver={driver}"

# Create SQLAlchemy engine
engine = create_engine(connection_string)

# Define your DataFrames and their corresponding table names
dfs = {
    'TM_02_LOC_04_FactLine': df_lines,
    'TM_02_LOC_01_Country': df_country,
    'TC_01_Demand_01_FGDemand': df_demand,
    'TC_02_FRT_01_Sec': df_Secondary_Freight,
    'TC_02_FRT_02_Pri': df_Primary_Freight,
    'TC_03_Factories_01_BOM': df_BOM,
    'TC_03_Factories_02_WorkingDays': df_calender,
    'TC_03_Factories_03_ShiftData': df_Factory_Operations,
    'TC_03_Factories_04_LineFGSKUData': df_Production,
    'TC_03_Factories_06_LabourSupply': df_LabourData,
    'TC_04_INV_01_FGOS': df_FG_M1OS,
    'TC_04_INV_02_RMOS': df_RM_M1OS,
    'TC_04_INV_04_RMInv': df_RM_Supply,
    'TM_01_Material_01_FGSKU': df_fgsku,
    'TM_01_Material_02_RMSKU': df_rmsku,
    'TM_01_Material_03_TrimSKU': df_trimming,
    'TM_02_LOC_02_Warehouse': df_warehouse,
    'TM_02_LOC_03_Factory': df_factory,
    'T_01_RM_transfer': df_rm_transfer,
    'T_01_FGfact_trimm':df_fgfact_trimm,
    'T_01_WH_FG':df_wh_fg,
    'T_01_RMTransfer_Restrictions':df_rmtransfer_restrictions,
    'T_01_Min_contract_labour':df_min_contract_labour,
    'T_01_Prep_line_FG':df_prep_line_fg,
    'T_01_Prep_line_capacity':df_prep_line_capacity,
    'TC_04_INV_06_FGCSInvNorm':df_TC_04_INV_06_FGCSInvNorm,
    'TC_04_INV_03_FGInv':df_TC_04_INV_03_FGInv,
    'TC_06_NegativeList_01_FGSupply':df_TC_06_NegativeList_01_FGSupply,
    'TC_06_NegativeList_02_RMSupply':df_TC_06_NegativeList_02_RMSupply,
    'TM_04_FIN_01_ICC':df_TM_04_FIN_01_ICC,
    'TM_03_CAL_01_Months':df_TM_03_CAL_01_Months,  
}

# Define datatypes for each column in each table
column_datatypes = {
    'TM_02_LOC_04_FactLine': {
        'FactCode': NVARCHAR(length=50),
        'Line': NVARCHAR(length=50),
    },
    'TM_02_LOC_01_Country': {
        'CountryCode': NVARCHAR(length=50),
        'CountryDesc': NVARCHAR(length=50),
    },
    'TC_01_Demand_01_FGDemand':{
        'FGSKUCode': NVARCHAR(length=50),
        'Country': NVARCHAR(length=50),
        'Month':INTEGER,
        'UM':NVARCHAR(length=50),
        'Demand':FLOAT,
    },
    'TC_02_FRT_01_Sec':{
        'Country':NVARCHAR(length=50),
        'WHCode':NVARCHAR(length=50),
        'FG Code':NVARCHAR(length=50),
        'Total Cost(FRT_CD)':FLOAT,
        'MaxShipment':FLOAT,
    },
    'TC_02_FRT_02_Pri':{
        'Origin':NVARCHAR(length=50),
        'WHCode':NVARCHAR(length=50),
        'FG Code':NVARCHAR(length=50),
        'Total Cost(FRT_CD)':FLOAT,
        'MaxShipment':FLOAT,
    },
    'TC_03_Factories_01_BOM':{
        'Bom version':NVARCHAR(length=50),
        'FactCode':NVARCHAR(length=50),
        'FG CODE':NVARCHAR(length=50),
        'FG DESC':NVARCHAR(length=250),
        'FG UOM':NVARCHAR(length=50),
        'RM CODE':NVARCHAR(length=50),
        'RM DESC':NVARCHAR(length=250),
        'RM UOM':NVARCHAR(length=50),
        'Consumption per unit of FG (KG per UOM)':FLOAT,
        'Cum Share taken':FLOAT,
    },
    'TC_03_Factories_02_WorkingDays':{
        'FactCode':NVARCHAR(length=50),
        'MonthNum':INTEGER,
        'HoursPerDay':FLOAT,
        'WorkingDays':FLOAT,
    },
    'TC_03_Factories_03_ShiftData':{
        'FactCode':NVARCHAR(length=50),
        'LineNum':NVARCHAR(length=50),
        'ShiftNum':INTEGER,
        'StandardHoursPerDay':FLOAT,
        'OT':FLOAT,
        'MaxVirtualShift':FLOAT,
        'CostVirtualShift':FLOAT,
        'CostRealShift':FLOAT,
        'CostOT':FLOAT,
    },
    'TC_03_Factories_04_LineFGSKUData':{
        'FactCode':NVARCHAR(length=50),
        'LineNum':NVARCHAR(length=50),
        'Unified code':NVARCHAR(length=50),
        'MOQ':FLOAT,
        'ThruputPerHour':FLOAT,
        'EnergyCostPerUnitBOM1':FLOAT,
        'EnergyCostPerUnitBOM2':FLOAT,
        'TeamSizeBOM1':FLOAT,
        'TeamSizeBOM2':FLOAT,
    },
    'TC_03_Factories_06_LabourSupply':{
        'FactCode':NVARCHAR(length=50),
        'MonthNum':INTEGER,
        'R_Days':FLOAT,
        'R_Count':FLOAT,
        'R_Min':FLOAT,
        'R_Max':FLOAT,
        'R_Hrs':FLOAT,
        'R_OT':FLOAT,
        'R_Wages':FLOAT,
        'R_OTWages':FLOAT,
        'C_Days':FLOAT,
        'C_Count':FLOAT,
        'C_OT':FLOAT,
        'C_OTWages':FLOAT,
        'RegCSMin':FLOAT,
        'RegCSMax':FLOAT,
        'CtrCSMin':FLOAT,
        'CtrCSMax':FLOAT,
        'RegMaxHire':FLOAT,
        'RegMaxFire':FLOAT,
        'CtrMaxHire':FLOAT,
        'CtrMaxFire':FLOAT,
    },
    'TC_04_INV_01_FGOS':{
        'FGSKUCode':NVARCHAR(length=50),
        'UM':NVARCHAR(length=50),
        'WH Code':NVARCHAR(length=50),
        'From NFC':FLOAT,
        'From GFC':FLOAT,
        'From KFC':FLOAT,
    },
    'TC_04_INV_02_RMOS':{
        'RM Code':NVARCHAR(length=50),
        'RM UOM':NVARCHAR(length=50),
        'FactCode':NVARCHAR(length=50),
        'RM-M1OS':FLOAT,
    },
    'TC_04_INV_04_RMInv':{
        'RM Code':NVARCHAR(length=50),
        'RMUOM':NVARCHAR(length=50),
        'FactCode':NVARCHAR(length=50),
        'CostPerUnit':FLOAT,
        'MaxOS':FLOAT,
        'MinOS':FLOAT,
        'Month':INTEGER,
        'MinCS':FLOAT,
        'MaxCS':FLOAT,
        'MaxSupLim':FLOAT,
        'RMStoreCost':FLOAT,
    },
    'TM_01_Material_01_FGSKU':{
        'FG Code':NVARCHAR(length=50),
        'FG Desc':NVARCHAR(length=250),
        'FG Cat':NVARCHAR(length=50),
        'FG UOM':NVARCHAR(length=50),
        'WtPerUnit':FLOAT,
        'Book Cost':FLOAT,
    },
    'TM_01_Material_02_RMSKU':{
        'RM Code':NVARCHAR(length=50),
        'RM UOM':NVARCHAR(length=50),
        'RM Desc':NVARCHAR(length=250),
    },
    'TM_01_Material_03_TrimSKU':{
        'TrimmingSKUCode':NVARCHAR(length=50),
        'TrimmingSKUDesc':NVARCHAR(length=250),
    },
    'TM_02_LOC_02_Warehouse':{
        'WHCode':NVARCHAR(length=50),
        'WHDesc':NVARCHAR(length=50),
        'WHLoc':NVARCHAR(length=50),
        'WHCountry':NVARCHAR(length=50),
        'PalletCapacity':FLOAT,
        'CostPerPalletPerMonth':FLOAT,
    },
    'TM_02_LOC_03_Factory':{
        'FactCode':NVARCHAR(length=50),
        'FactDesc':NVARCHAR(length=50),
        'FactLoc':NVARCHAR(length=50),
        'FactCountry':NVARCHAR(length=50),
    },
    'T_01_RM_transfer':{
        'OrigFact':NVARCHAR(length=50),
        'DestFact':NVARCHAR(length=50),
        'TSKUCode':NVARCHAR(length=50),
        'MthNum':INTEGER,
        'FrtRate/kg ':FLOAT,
    },
    'T_01_FGfact_trimm':{
        'ASSEMB_ITEM_CODE':NVARCHAR(length=50),
        'Fact Code':NVARCHAR(length=50),
        'TrimmingKGPerUnit':FLOAT,
    },
    'T_01_WH_FG':{
        'FG Code':NVARCHAR(length=50),
        'UOM':NVARCHAR(length=50),
        'WH Code':NVARCHAR(length=50),
        'CostperUOM':FLOAT,
    },
    'T_01_RMTransfer_Restrictions':{
        'OrigFact':NVARCHAR(length=50),
        'DestFact':NVARCHAR(length=50),
        'TSKUCode':NVARCHAR(length=50),
        'MaxLimit':INTEGER,
    },
    'T_01_Min_contract_labour':{
        'Factory':NVARCHAR(length=50),
        'MinContractLabourHoursPerMonth':INTEGER,
    },
    'T_01_Prep_line_FG':{
        'Factory':NVARCHAR(length=50),
        'Combination':NVARCHAR(length=50),
        'SKU Code':NVARCHAR(length=50),
    },
    'T_01_Prep_line_capacity':{
        'Factory':NVARCHAR(length=50),
        'Combination':NVARCHAR(length=50),
        'maxProdKg':INTEGER, 
    },
    'TC_04_INV_06_FGCSInvNorm':{
        'MinCS':FLOAT,
        'MaxCS':FLOAT,
        'MaxSupLim':FLOAT,
    }
}


tables_updated = []  # List to store the names of tables that were successfully updated

try:
    # Insert each DataFrame into its corresponding table
    for table_name, df in dfs.items():
        # Load your DataFrame for the current table
#         df = pd.read_csv(f'{table_name}.csv')  # Assuming your data is in CSV files
        
        # Get column datatypes for the current table
        table_column_datatypes = column_datatypes.get(table_name, {})
        
        # Transfer the DataFrame to the SQL Server table with specified datatypes
        df.to_sql(table_name, con=engine, if_exists='replace', index=False, dtype=table_column_datatypes)
        tables_updated.append(table_name)  # Add the table name to the list of updated tables

        
    # If no errors occurred, print all tables updated
    print("All tables updated successfully:")
    for table_name in tables_updated:
        print(table_name)

except Exception as e:
    print(f"An error occurred while inserting into table {table_name}: {e}")

finally:
    # Dispose the engine
    engine.dispose()
